extracted Contacts table to separate component, added Asc Column sort to all thead
